pyproject.toml
[project]
name = "linkedin-importer"
version = "0.1.0"
description = "LinkedIn Profile Importer - Import LinkedIn profile data to PostgreSQL database"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "asyncpg>=0.29.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
linkedin-importer = "linkedin_importer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "hypothesis>=6.0.0",
    "pytest-asyncio>=0.23.0",
]
